+++ src/main/java/xyz/dogboy/swp/event/CommonEventHandler.java
@@ -51,7 +51,7 @@ public class CommonEventHandler {
             ItemStack output = ((ItemBlockWoodenVariation) Registry.PIPE_ITEM).getWithBaseBlock(baseBlock);
             output.setCount(6);
 
-            return getRecipe(String.format("pipe_%s_%s_%d", baseBlock.getItem().getRegistryName().getNamespace(), baseBlock.getItem().getRegistryName().getNamespace(), baseBlock.getMetadata()), output,
+            return getRecipe(String.format("pipe_%s_%s_%d", baseBlock.getItem().getRegistryName().getNamespace(), baseBlock.getItem().getRegistryName().getPath(), baseBlock.getMetadata()), output,
                     baseBlockIngredient, glassIngredient, baseBlockIngredient, baseBlockIngredient, glassIngredient, baseBlockIngredient, baseBlockIngredient, glassIngredient, baseBlockIngredient);
         }).forEach(event.getRegistry()::register);
     }